requirements.txt
pre-commit
thrift
AWSIoTPythonSDK
Enum
pypubsub
lz4tools
mock
